Refactor color settings component in TS
